+++ api/src/models/structure.js
@@ -235,6 +235,12 @@ const Schema = new mongoose.Schema({
     lon: { type: Number },
     lat: { type: Number },
   },
+  addressVerified: {
+    type: String,
+    documentation: {
+      description: "L'adresse a été vérifiée",
+    },
+  },
 
   state: {
     type: String,